Add files via upload
Added CreatePictureComment.xlam - allows user to select a picture file and place in a comment on the selected cell.
DatePicker - Fixed leap year bug if a date before March 1, 1900 is chosen. User is no longer able to select a date before January 1, 1900. (Selecting December 31, 1900 is still possible, but it results in Excel displaying January 0, 1900).